created services folder to seperate data access logic
